Use common renderWithRouterAndTheme
