+++ frontend/tests/unit/components/Home.test.tsx
@@ -5,13 +5,13 @@ import { renderWithRouterAndTheme } from "tests/utils/renderWithRouterAndTheme";
 
 describe("Home", () => {
   it("renders title and description correctly", () => {
-    renderWithRouterAndTheme();
+    renderWithRouterAndTheme({ useRealHome: true });
     expect(screen.getByText("I18N_Title")).toBeInTheDocument();
     expect(screen.getByText("I18N_Description")).toBeInTheDocument();
   });
 
   it("renders statistics with correct labels and counts", () => {
-    renderWithRouterAndTheme();
+    renderWithRouterAndTheme({ useRealHome: true });
     expect(screen.getByText("I18N_Home_Stats_Productions")).toBeInTheDocument();
     expect(screen.getByText("I18N_Home_Stats_Events")).toBeInTheDocument();
     expect(screen.getByText("I18N_Home_Stats_Artists")).toBeInTheDocument();
@@ -19,7 +19,7 @@ describe("Home", () => {
   });
 
   it("renders buttons with correct labels and links", () => {
-    renderWithRouterAndTheme();
+    renderWithRouterAndTheme({ useRealHome: true });
     expect(
       screen.getByRole("link", { name: "I18N_Home_Button_Explore" })
     ).toBeInTheDocument();
@@ -30,21 +30,19 @@ describe("Home", () => {
 
   it("explorebutton navigates to archive page", async () => {
     const user = userEvent.setup();
-    renderWithRouterAndTheme();
+    renderWithRouterAndTheme({ useRealHome: true });
     const exploreButton = screen.getByTestId("home-button-I18N_Home_Button_Explore");
 
-    expect(exploreButton).toHaveAttribute("href", "/nl/archive");
     await user.click(exploreButton!);
-    expect(screen.getByText("I18N_Archive_Title")).toBeInTheDocument();
+    expect(screen.getByText("TEST_ARCHIVE_PAGE")).toBeInTheDocument();
   });
 
   it("historybutton navigates to history page", async () => {
     const user = userEvent.setup();
-    renderWithRouterAndTheme();
+    renderWithRouterAndTheme({ useRealHome: true });
     const historyButton = screen.getByTestId("home-button-I18N_Home_Button_History");
 
-    expect(historyButton).toHaveAttribute("href", "/nl/history");
     await user.click(historyButton!);
-    expect(screen.getByText("I18N_History_Title")).toBeInTheDocument();
+    expect(screen.getByText("TEST_HISTORY_PAGE")).toBeInTheDocument();
   });
 });
